Remove the top-level region option
The current implementation uses value set via the envvar AWS_DEFAULT_REGION
for sub-commands even if an explicit flag value is provided.

This commit also integrates get_region() into the 'region' option's
callback, so that senza first ensures a non-empty region value (gets
region from "~/.aws/config" if nothing is provided) and then validates
it against the available regions list.
